feat: STFT-076: Secureli ignore - ability to suppress issues - auto prompt user (#18)
[STFT-076: STFT-076: Secureli ignore - ability to suppress issues - auto
prompt user](https://slalom.atlassian.net/browse/STFT-76)

This PR adds a new feature which prompts users after a scan that has
resulted in a linter failure if they would like to add an ignore rule
for any of the detected failures.

If the user agrees to add an ignore rule, then secureli iterates through
each failure and asks the user if they want to add an ignore for that
failure, if so, then it also asks if the user wants to ignore the
failure for all files or a specific file.

To test this, make any change (and stage the file) that would cause a
linter failure. Some simple ones would be adding some white space at the
end of a line or removing the blank line at the end of a file. Any
linter failure should trigger the notification.

Note: New ignore rules are detected as an available upgrade, so when you
run your next scan after adding an ignore, you will want to confirm the
upgrade.