chore: add scaffolding for integration testing to CI pipeline and Dockerfile for local integration testing (#8)
this change adds a section to the pipeline that builds on a macos
container, installs homebrew, taps the secureli tap and installs
secureli. After this, next steps will be to make a local dev container
that does the same as well as add on to this test section with a step
that checks out an app and init's secureli into it.

I had some free time so I added a checkout for pip. The last section
clones the pip public repo, changes directory into it and runs `secureli
init`

This PR also adds a dockerfile for local test verification that mimics
the same steps outlined above only on a debian container.

There is a new section in the README that explains the docker
implementation

I also ran a `poetry update` while troubleshooting some problems so now
we have updated libraries in the project. Probably fine.