chore: secureli 91 add windows integration pipeline (#97)
Adds a build/test stage for windows into the pipeline. This is very
basic as the python tests are erroring right now. Commented out til we
figure that part out. This is so that ideally a windows dev can also
contribute.

Adds a job for windows integration testing - installs python, installs
pip, installs the secureli package via pip, clones a python repo, inits
secureli and runs a scan.

Cleans up the pipeline, removing an unneeded block that appears to be a
duplicate from when we separated out the homebrew deploy.

Removes the need for gh cli from the pypi dockerfile. GH CLI is not
required, only git for `clone`. This is in response to troubles on
windows with trying to add the gh cli repository key. We don't need it
anyway so it's a good change