chore: update project deps (#102)
Dependabot is not curerntly creating a PR because it can't parse the
pyproject.toml. It doesn't say what is the problem and a validator says
the file is fine. So I am running basic dep update to the poetry file to
see if that makes any kind of difference.

If it doesn't the next step is probably to make a separate bug and I
will poke Korey and maybe make an issue at the dependabot repo to
understand why it doesn't give us a better error. Other issues there
when they have trouble report actual line errors in the file and not
just "can't parse."

---------

Co-authored-by: Korey Earl <korey.earl@gmail.com>