fix: Correctly parsing %-encoded URLs from .git/config file (#331)
Resolves #261

The `configparser` module by default assumes that '%' characters
indicate the presence of a variable and will try to parse a config value
accordingly. Switching to "raw" mode prevents this behavior.

## Testing
Manually tested by changing my `.git/config` file to set the origin URL
to include `%20`, and was able to replicate the failure detailed in the
ticket.