Remove .pre-commit-config.yaml model in favor of existing classes
I found that the model I created already existed (in a slightly
different form). So I've updated the existing model to work for my use
case. Also, I updated the model to extend from BaseModel instead of
BaseSettings, since we probably don't want environment variables to
override the config in the file.