Force MacOS smoke test to run with python 3.11
And include commented-out code to enable python 3.12 support.
The python package dependency-injector doesn't yet support python 3.12.
There is a PR for enabling that support that is not yet merged.
I've opted to leave that commented-out because it will cause the
build to break once the PR branch it points to gets merged/deleted.