fix: Fix integration tests action (#364)
# Description

Attempting to fix our `Integration Tests` GH action (which is really
post-release smoke testing)

There are 3 failing jobs:
1. pulling from PyPI on Windows
a. This was failing because the scan returns a nonzero exit code. For
now, i told the script to expect a nonzero exit code.
b. In the future, we should have a "happy path" test case where no
issues are found, so we can validate a success (zero) exit code.
c. Additionally, we should use different exit codes to differentiate
between failures because of flagged issues versus internal errors. For
now, I have updated seCureLI to return a `3` when it flags issues, but
I'm not explicitly checking for that error here.
2. pulling from PyPI on MacOS
a. This was failing due to using python 3.12 in the default runner. I
fixed it by installing & using python 3.11 on the runner.
b. We can't yet support python 3.12 because one of our dependencies
doesn't yet support it. If we need it in a pinch, I've included
commented-out code for pointing to the PR branch that enables this
support.
3. pulling from homebrew on MacOS
a. This is failing because of our homebrew formula has an incorrect list
of dependencies. It will need to be fixed in a future PR.

# Testing

This execution shows tests passing (except for homebrew): https://github.com/slalombuild/secureli/actions/runs/7535989887

This closes #357