Move .pre-commit-config.yal into .secureli folder
Using a custom location for our pre-commit hooks allows us to
not destroy whatever existing configuration exists, opening the
door to interoperability with existing installations of pre-commit
and avoiding conflicts when e.g. we want to use a different set of hooks
(or different versions of hooks) that what was already configured.